requirements.txt
Pillow
pyfiglet
keyboard
shutil